Move tests and update imports
Relocate tests closer to the code being tested and update imports.

Ticket: AT-7050
