+++ api/provision/start-provision-job.test.ts
@@ -1,9 +1,9 @@
 import { Context } from "aws-lambda";
-import { handler } from "../api/provision/start-provision-job";
-import { CloudServiceProvider, Network } from "../models/cloud-service-providers";
-import { ProvisionRequestType } from "../models/provisioning-jobs";
-import { ApiSuccessResponse, ValidationErrorResponse } from "../utils/response";
-import { sfnClient } from "../utils/aws-sdk/step-functions";
+import { handler } from "./start-provision-job";
+import { CloudServiceProvider, Network } from "../../models/cloud-service-providers";
+import { ProvisionRequestType } from "../../models/provisioning-jobs";
+import { ApiSuccessResponse, ValidationErrorResponse } from "../../utils/response";
+import { sfnClient } from "../../utils/aws-sdk/step-functions";
 import { mockClient } from "aws-sdk-client-mock";
 
 const sfnMock = mockClient(sfnClient);
