+++ lib/constructs/apigateway.test.ts
@@ -1,8 +1,8 @@
 import * as cdk from "aws-cdk-lib";
 import * as ec2 from "aws-cdk-lib/aws-ec2";
 import * as iam from "aws-cdk-lib/aws-iam";
-import { Match, Template } from "aws-cdk-lib/assertions";
-import { AtatRestApi, RestApiVpcConfiguration } from "../lib/constructs/apigateway";
+import { Template } from "aws-cdk-lib/assertions";
+import { AtatRestApi, RestApiVpcConfiguration } from "./apigateway";
 
 describe("ATAT apigateway construct creation", () => {
   test("Ensure AtatRestApi is created", async () => {
